+++ packages/wrangler/src/api/startDevWorker/ConfigController.ts
@@ -1,0 +1,42 @@
+import { EventEmitter } from "node:events";
+// import { readFileSync } from "../../parse";
+import { throwNotImplementedError } from "./utils";
+import type { ConfigUpdateEvent, ErrorEvent } from "./events";
+import type { StartDevWorkerOptions } from "./types";
+
+export class ConfigController extends EventEmitter {
+	config?: StartDevWorkerOptions;
+
+	setOptions(_: StartDevWorkerOptions) {
+		throwNotImplementedError(this.setOptions.name, this.constructor.name);
+	}
+	updateOptions(_: Partial<StartDevWorkerOptions>) {
+		throwNotImplementedError(this.updateOptions.name, this.constructor.name);
+	}
+
+	// ******************
+	//   Event Handlers
+	// ******************
+
+	async teardown() {
+		throwNotImplementedError(this.teardown.name, this.constructor.name);
+	}
+
+	// *********************
+	//   Event Dispatchers
+	// *********************
+
+	emitConfigUpdateEvent(data: ConfigUpdateEvent) {
+		this.emit("configUpdate", data);
+	}
+
+	// *********************
+	//   Event Subscribers
+	// *********************
+
+	on(event: "configUpdate", listener: (_: ConfigUpdateEvent) => void): this;
+	// @ts-expect-error Missing overload implementation (only need the signature types, base implementation is fine)
+	on(event: "error", listener: (_: ErrorEvent) => void): this;
+	// @ts-expect-error Missing initialisation (only need the signature types, base implementation is fine)
+	once: typeof this.on;
+}
